+++ core/NumberFormatter.php
@@ -42,6 +42,22 @@ class NumberFormatter
     }
 
     /**
+     * Parses the given pattern and returns patterns for positive and negative numbers
+     *
+     * @param string $pattern
+     * @return array
+     */
+    protected function parsePattern($pattern)
+    {
+        $patterns = explode(';', $pattern);
+        if (!isset($patterns[1])) {
+            // No explicit negative pattern was provided, construct it.
+            $patterns[1] = '-' . $patterns[0];
+        }
+        return $patterns;
+    }
+
+    /**
      * Formats a given number or percent value (if $value starts or ends with a %)
      *
      * @param string|int|float $value
@@ -166,22 +182,6 @@ class NumberFormatter
     }
 
     /**
-     * Parses the given pattern and returns patterns for positive and negative numbers
-     *
-     * @param string $pattern
-     * @return array
-     */
-    protected function parsePattern($pattern)
-    {
-        $patterns = explode(';', $pattern);
-        if (!isset($patterns[1])) {
-            // No explicit negative pattern was provided, construct it.
-            $patterns[1] = '-' . $patterns[0];
-        }
-        return $patterns;
-    }
-
-    /**
      * Formats the given number with the given pattern
      *
      * @param string $pattern
@@ -236,10 +236,10 @@ class NumberFormatter
             // Reconstruct the major digits.
             $majorDigits = implode(',', $groups);
         }
-        if ($minimumFractionDigits < $maximumFractionDigits) {
+        if ($minimumFractionDigits <= $maximumFractionDigits) {
             // Strip any trailing zeroes.
             $minorDigits = rtrim($minorDigits, '0');
-            if (strlen($minorDigits) < $minimumFractionDigits) {
+            if (strlen($minorDigits) && strlen($minorDigits) < $minimumFractionDigits) {
                 // Now there are too few digits, re-add trailing zeroes
                 // until the desired length is reached.
                 $neededZeroes = $minimumFractionDigits - strlen($minorDigits);
@@ -297,5 +297,11 @@ class NumberFormatter
     public static function getInstance()
     {
         return StaticContainer::get('Piwik\NumberFormatter');
+    }
+
+    public function clearCache()
+    {
+        $this->patterns = [];
+        $this->symbols = [];
     }
 }