+++ core/Plugin/ViewDataTable.php
@@ -239,6 +239,9 @@ abstract class ViewDataTable implements ViewInterface
          * Triggered during {@link ViewDataTable} construction. Subscribers should customize
          * the view based on the report that is being displayed.
          *
+         * This event is triggered before view configuration properties are overwritten by saved settings or request
+         * parameters. Use this to define default values.
+         *
          * Plugins that define their own reports must subscribe to this event in order to
          * specify how the Piwik UI should display the report.
          *
@@ -275,6 +278,31 @@ abstract class ViewDataTable implements ViewInterface
 
         $this->overrideViewPropertiesWithParams($overrideParams);
         $this->overrideViewPropertiesWithQueryParams();
+
+        /**
+         * Triggered after {@link ViewDataTable} construction. Subscribers should customize
+         * the view based on the report that is being displayed.
+         *
+         * This event is triggered after all view configuration values have been overwritten by saved settings or
+         * request parameters. Use this if you need to work with the final configuration values.
+         *
+         * Plugins that define their own reports can subscribe to this event in order to
+         * specify how the Piwik UI should display the report.
+         *
+         * **Example**
+         *
+         *     // event handler
+         *     public function configureViewDataTableEnd(ViewDataTable $view)
+         *     {
+         *         if ($view->requestConfig->apiMethodToRequestDataTable == 'VisitTime.getVisitInformationPerServerTime'
+         *             && $view->requestConfig->flat == 1) {
+         *                 $view->config->show_header_message = 'You are viewing this report flattened';
+         *         }
+         *     }
+         *
+         * @param ViewDataTable $view The instance to configure.
+         */
+        Piwik::postEvent('ViewDataTable.configure.end', array($this));
     }
 
     private function assignRelatedReportsTitle()